Split into a frontend/backend communicating over TLS/Websocket
By splitting the system_monitor into a frontend/backend
it is possible to output the collected data in the frontend
to another host where the backend (and Postgres) run.

The communication is done over TLS and a Websocket,
using both Client- and Server certificates to protect
the data being sent.

Adding the 'nosup3' rebar3 profile in order to exclude
the use of the supervisor3 application, in favour of
the standard OTP supervisor application.

Made the supervisor restart intensity configurable.

Documented in CONSUMER_PRODUCER.md.