Enable custom metrics collectors
By setting the application env vars: custom_metrics
it is possible to hook in custom metric collectors.

Adding the system_monitor_memory_metric module that
collect system Total and Free memory as well as the
Allocated and Used memory of the target system, and
store it into the memory DB table.

Collecting data for the ERTS allocators and store
them into the ertsalloc DB table.

Note: The corresponding Postgres tables must be created,
e.g similar to how it is done in the: 40-memory.sh
and 41-ertsalloc.sh scripts.

Also, documenting the status_check and external_monitoring
app env variables.