Fixing the Memory dashboard
I had created the memory.json by exporting it from Grafana,
but apparently the way "datasource" was expressed was wrong.
So after changing it to look like the other dashboard files
it worked to import it to a fresh Grafana instance.

Perhaps the initial json should have been exported as:
'Export for sharing externally '?